Separate AllProducts into 2 part, AllProductsCustomer for view all products, AllProduts for admin to manage products
