+++ src/App.jsx
@@ -5,20 +5,19 @@ import Login from "./pages/Login.jsx";
 import RegularProducts from "./pages/RegularProducts.jsx";
 import BestSellingProducts from "./pages/BestSellingProducts.jsx";
 import AllProducts from "./pages/AllProducts.jsx";
-import SignUp from "./pages/SignUp.jsx"; // Thêm SignUp
-import AuthGuard from "./guards/AuthGuard.jsx"; // Import AuthGuard
+import AllProductsCustomer from "./pages/AllProductsCustomer.jsx";
+import SignUp from "./pages/SignUp.jsx";
+import AuthGuard from "./guards/AuthGuard.jsx";
 import { ToastContainer } from "react-toastify";
 
 function App() {
   return (
     <div>
       <Routes>
-        <Route path="/" element={<Home />} /> {/* Trang công khai */}
-        <Route path="/login" element={<Login />} /> {/* Trang công khai */}
-        <Route path="/signup" element={<SignUp />} /> {/* Trang công khai */}
-        <Route path="/products/all" element={<AllProducts />} />{" "}
-        {/* Trang công khai */}
-        {/* Route bảo vệ bằng AuthGuard */}
+        <Route path="/" element={<Home />} />
+        <Route path="/login" element={<Login />} />
+        <Route path="/signup" element={<SignUp />} />
+        <Route path="/products/all" element={<AllProductsCustomer />} />
         <Route element={<AuthGuard />}>
           <Route path="/card-management" element={<CardManagement />} />
           <Route path="/products/regular" element={<RegularProducts />} />
@@ -26,6 +25,7 @@ function App() {
             path="/products/best-selling"
             element={<BestSellingProducts />}
           />
+          <Route path="/products/all-management" element={<AllProducts />} />
         </Route>
       </Routes>
       <ToastContainer position="bottom-right" autoClose={3000} closeOnClick />
